refactor(views): move views into component directories
